Moved the FnCtxt struct to its own file
